requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pymysql==1.1.0
google-genai==1.50.0
python-dotenv==1.0.0
PyMuPDF==1.26.7
pandas
gunicorn==21.2.0
Flask-WTF==1.2.1
Flask-Limiter==3.5.0
bleach==6.1.0
Waitress==3.0.0
Flask-Compress==1.14
boto3>=1.35.0
